Split makefile generation into smaller files
- makefile for the core model types and render function.
- makefile-builders for the functions to build specific makefiles.
